fix: Don't disable nonexistant fedora-cisco-openh264

+++ system_files/nvidia/tmp/nvidia-post-install.sh
@@ -8,8 +8,12 @@ mv /etc/nvidia-container-runtime/config.toml{,.orig}
 cp /etc/nvidia-container-runtime/config{-rootless,}.toml
 
 semodule --verbose --install /usr/share/selinux/packages/nvidia-container.pp
-ln -s /usr/bin/ld.bfd /etc/alternatives/ld
-ln -s /etc/alternatives/ld /usr/bin/ld
+if [[ ! -f /etc/alternatives/ld ]]; then
+    ln -s /usr/bin/ld.bfd /etc/alternatives/ld
+fi
+if [[ ! -f /usr/bin/ld ]]; then
+    ln -s /etc/alternatives/ld /usr/bin/ld
+fi
 
 if [[ "${IMAGE_NAME}" == "sericea" ]]; then
     mv /etc/sway/environment{,.orig}
